Reduce CPU count to 1
My recent experience is that reducing the Virtual CPU setting speeds up performance considerably. This appears to be supported by other articles out there on the interwebs. I'm proposing it as a change to your template since I have used it as a reference point a lot. Test the change and see if you see a performance improvement too.

For more information see:
https://ruin.io/2014/benchmarking-virtualbox-multiple-core-performance/
https://github.com/roots/trellis/issues/410